fix: fix warning double to float

+++ include/pland/Calculate.h
@@ -17,8 +17,8 @@ public:
 
     int eval(string code);
 
-    static int calculateDiscountPrice(int originalPrice, float discountRate); // 计算折扣后的价格
-    static int calculateRefundsPrice(int originalPrice, float refundRate);    // 计算退款金额
+    static int calculateDiscountPrice(int originalPrice, double discountRate); // 计算折扣后的价格
+    static int calculateRefundsPrice(int originalPrice, double refundRate);    // 计算退款金额
 };
